% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_map.R
\name{base_map}
\alias{base_map}
\title{Builds basic leaflet map based on study area (shape)}
\usage{
base_map(shape_name)
}
\arguments{
\item{shape}{the study area which should be displayed by the map}
}
\value{
a leaflet map
}
\description{
Builds basic leaflet map based on study area (shape)
}
\examples{
data(confluence)
base_map("confluence")
}
