% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_osmdata.R
\name{clear_osmdata}
\alias{clear_osmdata}
\title{Clears the local {dir_name}/{shape_name}/{key}_{value}.RDS file if it exists}
\usage{
clear_osmdata(key, value, shape_name, dir_name = "data/osmdata")
}
\arguments{
\item{key}{OSM key}

\item{value}{OSM value}

\item{shape_name}{shape (sf object) or name of a shape (inside quotes), corresponding to the geographical area inside which to look for key-value occurrences}

\item{dir_name}{the directory under which returned data is saved. Defaults to "data/osmdata".}
}
\description{
Clears the local {dir_name}/{shape_name}/{key}_{value}.RDS file if it exists
}
\examples{
bridges_confluence=get_osmdata("man_made","bridge","confluence", save=TRUE)
file.exists("data/osmdata/confluence/man_made-bridge.RDS")
clear_osmdata("man_made","bridge","confluence")
file.exists("data/osmdata/confluence/man_made-bridge.RDS")
}
