% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_examples.R
\docType{data}
\name{condrieu}
\alias{condrieu}
\title{confluence}
\format{
An sf object corresponding to a rectangle
}
\usage{
condrieu
}
\description{
A bounding shape of about 20 squared kilometers corresponding to the area around Condrieu, France.
}
\keyword{datasets}
