% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include-some-data-examples-in-your-package.R,
%   R/shape_examples.R
\docType{data}
\name{confluence}
\alias{confluence}
\title{confluence}
\format{
An sf object corresponding to a rectangle

An sf object corresponding to a rectangle
}
\usage{
confluence

confluence
}
\description{
A bounding shape corresponding to the confluence of the Rhone and Saone rivers in Lyon, France.

A bounding shape of about 5.5 squared kilometers corresponding to the confluence of the Rhone and Saone rivers in Lyon, France.
}
\keyword{datasets}
