% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_osmdata_item.R
\name{describe_osmdata_item}
\alias{describe_osmdata_item}
\title{Describe the content of an osmdata directory for a particular key-value association}
\usage{
describe_osmdata_item(directory, key, value)
}
\arguments{
\item{directory}{the directory}

\item{key}{the osm key to consider}

\item{value}{the osm value to consider}
}
\value{
a table with the number, geometry type, and other characteristics of osm elements
}
\description{
Describe the content of an osmdata directory for a particular key-value association
}
