% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detail_element.R
\name{detail_element}
\alias{detail_element}
\title{Creates an html column detailing information regarding OSM items}
\usage{
detail_element(gosmdata)
}
\arguments{
\item{gosmdata}{the gosmdata element you wish to detail}
}
\value{
gosmdata with additional detail column
}
\description{
Creates an html column detailing information regarding OSM items
}
\examples{
g_parks=get_osmdata("leisure","park","confluence") \%>\% 
  pick_osmdata(geom="osm_polygons") \%>\% 
  detail_element()
g_parks
}
