% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch_osmdata_in_zones.R
\name{dispatch_osmdata_in_zones}
\alias{dispatch_osmdata_in_zones}
\title{intersect the OSM elements shapefile with a sf object defining zones.}
\usage{
dispatch_osmdata_in_zones(
  osm_shape_path,
  zones_shape,
  from = "data/osmdata",
  to = "data/osmdata_trimmed"
)
}
\arguments{
\item{osm_shape_path}{the path to a saved OSM elements shapefile (saved by save_osmdata() function)}

\item{zones_shape}{the shapes of zones in which to dispatch the osm elements (each OSM element will be associated to one or several zones)}

\item{from}{the name of the directory containing the original osm data shapefiles}

\item{to}{the name of the directory which will contain the resulting shapefiles}
}
\description{
intersect the OSM elements shapefile with a sf object defining zones.
}
