% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osmdata.R
\name{get_osmdata}
\alias{get_osmdata}
\title{collect OSM data corresponding to a key-value inside a bounding shape. If a file already exists corresponding to "{dir_name}/{shape_name}\emph{{key}}{value}, the data is not retrieved through a new query but through reading this file.}
\usage{
get_osmdata(keyvalue = NA, type = "osm_polygons", citycode, conn)
}
\arguments{
\item{keyvalue}{OSM keyvalue}

\item{type}{the type of OSM geometry (defaults to "osm_polygons")}

\item{citycode}{the citycode}

\item{conn}{a connection to database}
}
\value{
a list with elements key, value, shape_name and result which is an osmdata_sf object
}
\description{
collect OSM data corresponding to a key-value inside a bounding shape. If a file already exists corresponding to "{dir_name}/{shape_name}\emph{{key}}{value}, the data is not retrieved through a new query but through reading this file.
}
\examples{
citycode=1235681_21899
get_osmdata(keyvalue="landuse-farmland",type="osm_polygons",citycode="1235681_21899",conn=sandbox_conn())
}
