% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_examples.R
\docType{data}
\name{lyon}
\alias{lyon}
\title{lyon}
\format{
An sf object corresponding to a rectangle
}
\usage{
lyon
}
\description{
A bounding shape of about 460 squared kilometers corresponding to the area around Lyon, France.
}
\keyword{datasets}
