% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_osmdata.R
\name{pick_osmdata}
\alias{pick_osmdata}
\title{Pick the information corresponding to a certain geometry}
\usage{
pick_osmdata(osmdata, geom = "osm_points", label = NA)
}
\arguments{
\item{osmdata}{the result returned by get_osmdata}

\item{geom}{the type of geometry. Can be for instance "osm_points","osm_lines","osm_polygons"}

\item{label}{the label letting the user know the type of object}
}
\value{
an sf object
}
\description{
Pick the information corresponding to a certain geometry
}
\examples{
bridges_confluence=get_osmdata("man_made","bridge",confluence) 
plg_bridges_confluence=pick_osmdata(bridges_confluence,
                                    geom="osm_polygons")
water_confluence=get_osmdata("natural","water",confluence)
plg_water_confluence=pick_osmdata(water_confluence,
                                  geom="osm_polygons")
}
