% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_osmdata.R
\name{save_osmdata}
\alias{save_osmdata}
\title{collect OSM data corresponding to a key-value inside a bounding shape. If a file already exists corresponding to "{dir_name}/{shape_name}\emph{{key}}{value}, the data is not retrieved through a new query but through reading this file.}
\usage{
save_osmdata(key, value, shape_name, dir_name = "data/osmdata")
}
\arguments{
\item{key}{OSM key}

\item{value}{OSM value}

\item{shape_name}{shape (sf object) or name of a shape (inside quotes), or path to a shapefile (inside quotes and corresponding to an existing path) corresponding to the geographical area inside which to look for key-value occurrences}

\item{dir_name}{the directory under which returned data is saved. Defaults to "data/osmdata".}
}
\value{
a list with elements key, value, shape_name and result which is an osmdata_sf object
}
\description{
collect OSM data corresponding to a key-value inside a bounding shape. If a file already exists corresponding to "{dir_name}/{shape_name}\emph{{key}}{value}, the data is not retrieved through a new query but through reading this file.
}
\examples{
data(confluence)
save_osmdata("man_made","bridge",confluence)
drinking_water_confluence=save_osmdata("amenity","drinking_water",confluence)
data(condrieu)
landuse_vineyard_condrieu=save_osmdata("landuse","vineyard",condrieu)
data(lyon)
landuse_port_lyon=save_osmdata("landuse","port",lyon)
}
