% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include-some-data-examples-in-your-package.R
\docType{data}
\name{confluence}
\alias{confluence}
\title{confluence}
\format{
An sf object corresponding to a bounding box
}
\usage{
confluence
}
\description{
A bounding shape corresponding to the confluence of the Rhone and Saone rivers in Lyon, France.
}
\keyword{datasets}
