% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osmdata.R
\name{get_osmdata}
\alias{get_osmdata}
\title{collect OSM data corresponding to a key-value inside a bounding shape. If a file already exists corresponding to "{dir_name}/{shape_name}\emph{{key}}{value}, the data is not retrieved through a new query but through reading this file.}
\usage{
get_osmdata(
  key,
  value,
  shape_name,
  return = TRUE,
  save = FALSE,
  dir_name = "data/osmdata"
)
}
\arguments{
\item{key}{OSM key}

\item{value}{OSM value}

\item{shape_name}{shape (sf object) or name of a shape (inside quotes), corresponding to the geographical area inside which to look for key-value occurrences}

\item{return}{whether the object should be returned (could be FALSE when save is TRUE)}

\item{save}{if TRUE, the data is saved locally in directory dir_name, with name {key}_{value}.RDS. If the file already exists nothing is done. Defaults to FALSE. If}

\item{dir_name}{the directory under which returned data is saved. Defaults to "data/osmdata".}
}
\value{
a list with elements key, value, shape_name and result which is an osmdata_sf object
}
\description{
collect OSM data corresponding to a key-value inside a bounding shape. If a file already exists corresponding to "{dir_name}/{shape_name}\emph{{key}}{value}, the data is not retrieved through a new query but through reading this file.
}
\examples{
data(confluence)
bridges_confluence=get_osmdata("man_made","bridge","confluence") 
bridges_confluence=get_osmdata("man_made","bridge","confluence", save=TRUE)
drinking_water_confluence=get_osmdata("amenity","drinking_water","confluence",save=TRUE)
landuse_vineyard_condrieu=get_osmdata("landuse","vineyard","condrieu")
landuse_port_lyon=get_osmdata("landuse","port","lyon")
}
